backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
email-validator==2.1.0
apscheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymysql==1.1.0
alembic==1.13.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
user-agents==2.2.0
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.2.0
xlsxwriter==3.2.9
